requirements.txt
absl-py==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
attrs==22.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
cachelib==0.10.2
cachetools==5.2.0
camelCasing==0.1.3
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1670575068857/work
configobj==5.0.6
contourpy==1.0.6
cycler==0.10.0
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dulwich==0.20.50
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
everett==3.1.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
flatbuffers==23.1.21
fonttools==4.38.0
furl==2.1.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
glob2==0.7
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h5py==3.8.0
idna==2.10
importlib-metadata==5.2.0
ipykernel @ file:///D:/bld/ipykernel_1671555400921/work
ipython @ file:///D:/bld/ipython_1669904345744/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1670253809910/work
jupyter_core @ file:///D:/bld/jupyter_core_1669776347344/work
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1670530880680/work
pandas==1.5.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathlib2==2.3.7.post1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1670613602756/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1670414775770/work
protobuf==3.19.6
psutil==5.9.4
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
PyJWT==2.4.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==0.21.0
pytz==2022.7
pywin32==227
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roboflow==0.2.21
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
semantic-version==2.10.0
sentry-sdk==1.13.0
simplejson==3.18.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sounddevice==0.4.5
SQLAlchemy==2.0.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tflite-support-nightly==0.4.4.dev20221106
thop==0.1.1.post2209072238
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1671473371426/work
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
websocket-client==1.3.3
Werkzeug==2.2.2
wget==3.2
wincertstore==0.2
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.11.0
